pyproject.toml
[project]
name = "short-video-remove-watermark-admin-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
]
